Added Guard and Noble, updated gitignore
Set up the guard and noble's prefabs so they both have colliders and
tags (they don't need to do anything else).

Updated gitignore to actually ignore the right directories.
Because I had the project set up as i did (two unity projects in one
repository), and git assumed the unity project would BE the repository,
it didn't ignore any of the directories it was supposed to.

So now I've updated the git to ignore the proper directories, and if
I've fucked it up this time then fuck it.